Prove: rm module
